Now whenever a value is selected in a combo box, it updates the door and ironmongery as appropriate.
TODO:
- MOST IMPORTANTLY need to get the textfields to update the objects,
  then the custom second panel ones
- clear opening must be relative to leaf size
- partMCompliant combobox is to be based on the entrance level
- need to discus how the structural opening is to be done, such as if it
  is to be a combobox or a text field (also looking likely to have to
  use subclass of the Dimension class as its toString() method is
  annoying

If anyone looks through this code and wonders why the setters have both
Object and String/YesNoOptions parameter variants: basically, the
comoboxes used in this app are sometimes or type String and sometimes
YesNoOptions. To get around this I used generics and wild cards ?.
However when using a setter method reference, it means that the compiler
defaults to Object, so I have to catch this. The alternatives are either
to convert them all to String and then only have additional String
variant setters for the YesNoOptions setters; or to add a different
method to attach combo box handlers, one where the type parameter is a
String and another for where it is YesNoOptions. Both are probably more
effective and efficient and I may refactor in the future, but currently
this is the sitation.